Initial updates for Emacs 24.1
As part of the project to make it more fun to do Coffee Script in
Emacs I've made it compile ``cleanly'' on Emacs 24.1.

Since it relies on a cl package currently I had to ignore those
errors, although I'd like to ditch the dependence on `some' if it's
not really needed.

Also, started a ChangeLog and added a .gitignore.